css done

+++ db/migrate/20200503101037_create_trainings.rb
@@ -3,7 +3,7 @@ class CreateTrainings < ActiveRecord::Migration[5.2]
     create_table :trainings do |t|
       t.date :date
       t.string :content
-      t.float :time
+      t.integer :time
       t.references :category, foreign_key: true
 
       t.timestamps
